Edited out previous attempt
cut out the extra code used in my first iteration. Also moved the PATCH request to its own route in the API for future paths that may be necessary, such as GET for pulling all users on an adventure.
