+++ src/app/services/userService.tsx
@@ -31,14 +31,3 @@ export const updateUserProfile = async (name: string, password: string, userId: 
         return response.json(); 
     }
 };
-
-export const getUserByEmail = async (email: string) => {
-    const response = await fetch(`http://localhost:3000/api/users`, {
-        method: 'GET',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ email })
-    });
-    if (!response.ok) throw new Error('Error fetching user by email');
-    const data = await response.json();
-    return data.user;
-}
